Do not double-log response messages
In some cases, a consumer wants to do a PUT, POST, ect, but does not
care about what comes back. One way to do this is to provide a
`deserializeResponseContentAsync` that is more or less just the
following.

```cs
// ...
response => Task.FromResult(response)
// ...
```

The trouble with this, however, is that the default implementation of
`IHttpRequestSender` logs the "response content" after it is
deserialized, no matter what. So this adds a check to avoid logging our
"response content" if it is an HttpResponseMessage which is equal to the
actual response.